updated dvae to match floatX

+++ flows/flows/config.py
@@ -1,4 +1,4 @@
 import tensorflow as tf
 
-floatX = 'float32'
+floatX = 'float64'
 phase = tf.placeholder_with_default(True, shape=(), name='learning_phase')
